requirements.txt
selenium
webdriver_manager
bs4
pyaml
lxml
